+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Tab bar uses darkened background with visible divider lines between tabs
+- Tab bar has 1px bottom border separating it from pane content
+
 ## [0.8.1] - 2026-03-06
 
 ### Fixed